Predict an extremal random forest (ERF)

Predicts ... !!! write description

# S3 method for erf
predict(object, newdata = NULL, quantiles = c(0.95, 0.99), ...)

Arguments

object

Fitted "erf" object.

newdata

Numeric matrix (!!!) or data.frame of test predictor values at which predictions are to be made. If NULL, predictions are made on the training data instead. !!! Talk about out-of-bag predictions? For further information see grf::quantile_forest(). Default is NULL.

quantiles

Numeric vector of quantile levels at which estimates are to be made. Default is c(0.95, 0.99).

...

Dots. Currently ignored.

Value

Numeric matrix with predictions for each test point (rows) at the desired quantile levels (columns).

Details

!!! write details

References

There are no references for Rd macro \insertAllCites on this help page.

Examples

"!!! add examples"
#> [1] "!!! add examples"