Predict with a quantile forest

Gets estimates of the conditional quantiles of Y given X using a trained forest.

# S3 method for quantile_forest
predict(object, newdata = NULL, quantiles = NULL, num.threads = NULL, ...)

Arguments

object
The trained forest.

newdata
Points at which predictions should be made. If NULL, makes out-of-bag predictions on the training set instead (i.e., provides predictions at Xi using only trees that did not use the i-th training example). Note that this matrix should have the number of columns as the training matrix, and that the columns must appear in the same order.

quantiles
Vector of quantiles at which estimates are required. If NULL, the quantiles used to train the forest is used. Default is NULL.

num.threads
Number of threads used in training. If set to NULL, the software automatically selects an appropriate amount.

...
Additional arguments (currently ignored).

Value

Predictions at each test point for each desired quantile.

Examples

# \donttest{
# Train a quantile forest.
n <- 50
p <- 10
X <- matrix(rnorm(n * p), n, p)
Y <- X[, 1] * rnorm(n)
q.forest <- quantile_forest(X, Y, quantiles = c(0.1, 0.5, 0.9))

#> Error in quantile_forest(X, Y, quantiles = c(0.1, 0.5, 0.9)): could not find function "quantile_forest"

# Predict on out-of-bag training samples.
q.pred <- predict(q.forest)

#> Error in predict(q.forest): object 'q.forest' not found

# Predict using the forest.
X.test <- matrix(0, 101, p)
X.test[, 1] <- seq(-2, 2, length.out = 101)
q.pred <- predict(q.forest, X.test)

#> Error in predict(q.forest, X.test): object 'q.forest' not found

# }