Extreme predictions with a quantile forest

Gets estimates of the extrapolated conditional quantiles of Y given X using a trained forest.

predict_erf(
  object,
  quantiles,
  threshold = 0.8,
  newdata = NULL,
  model_assessment = FALSE,
  Y.test = NULL,
  out_of_bag = FALSE
)

Arguments

object
Quantile forest object. The trained forest.

quantiles
Numeric vector (0, 1). Extreme quantiles at which estimates are required.

threshold
Numeric (0, 1). Intermediate quantile used to compute thresholds \(t(x_i)\) for the GPD. Default is 0.8. Note that threshold < quantiles.

newdata
Numeric matrix. Rows contain observation at which predictions should be made. If NULL, makes out-of-bag predictions on the training set instead (i.e., provides predictions at Xi using only trees that did not use the i-th training example). Note that this matrix should have the number of columns as the training matrix, and that the columns must appear in the same order. This argument is needed if model_assessment = TRUE. Default is NULL.

model_assessment
Boolean. Assess GPD fit with QQ-plot against exponential distribution? Default is FALSE.

Y.test
Numeric vector. Responses for the test set, needed if model_assessment = TRUE. Default is NULL.

out_of_bag
Boolean. Use out-of-bag observations to compute thresholds \(t(x_i)\) for the GPD? Default is FALSE.

Value

Predictions at each test point for each desired quantile.