Extreme predictions with a quantile forest

Predicts high conditional quantiles of Y given X using a quantile forest fitted using quantile_forest.

predict_erf(
  object,
  quantiles,
  threshold = 0.8,
  newdata = NULL,
  model_assessment = FALSE,
  Y.test = NULL,
  out_of_bag = FALSE,
  param_est = c("ML", "Hill"),
  lambda = 0
)

Arguments

object
Quantile forest object. The trained forest obtained with quantile_forest.

quantiles
Numeric vector (0, 1). Extreme quantiles at which estimates are required.

threshold
Numeric (0, 1). Intermediate quantile used to compute thresholds \(t(x_i)\) for the GPD. Note that threshold < quantiles. Default is 0.8.

newdata
Numeric matrix. Rows contain observation at which predictions should be made. If NULL, makes out-of-bag predictions on the training set instead (i.e., provides predictions at Xi using only trees that did not use the i-th training example). Note that this matrix should have the number of columns as the training matrix, and that the columns must appear in the same order. This argument is needed if model_assessment = TRUE. Default is NULL.

model_assessment
Boolean. Assess GPD fit with QQ-plot against exponential distribution? Default is FALSE.

Y.test
Numeric vector. Responses for the test set, needed if model_assessment = TRUE. Default is NULL.

out_of_bag
Boolean. Use out-of-bag observations to compute thresholds \(t(x_i)\) for the GPD? Default is FALSE.

param_est
Character. One of "ML" and "Hill". Specifies the method to fit the shape and scale parameter of the GDP distribution. "ML" is the weighted maximum likelihood approach proposed by . "Hill" is the weighted Hill's estimator approach proposed by .

lambda
Numeric (>= 0). Penalty coefficient for the \(\xi\) parameter in the weighted log-likelihood. Default is 0.

Value

Named list. The list is made of:

predictions --- Numeric matrix. Predictions at each test point (on rows) for each desired quantile (on columns).

pars --- Numeric matrix. Estimated parameters at each test point (on rows). The columns contain the estimates for the \(\sigma\) and \(\xi\) parameter, respectively.

threshold --- Numeric matrix. Intermediate thresholds at each test point (on rows), estimated using quantile_forest.

lambda --- Numeric. Penalty coefficient for the \(\xi\) parameter in the weighted log-likelihood.

plot (if model_assessment = TRUE). QQ-plot for model assessment.